chore: isolate sendZammadMail function
