+++ src/connectors/sendinblue.ts
@@ -1,22 +1,18 @@
 import axios, { AxiosError, AxiosResponse } from "axios";
 import { chain, isEmpty } from "lodash";
 import path from "path";
-import {
-  DO_NOT_SEND_MAIL,
-  SENDINBLUE_API_KEY,
-  ZAMMAD_URL,
-} from "../config/env";
+import { DO_NOT_SEND_MAIL, SENDINBLUE_API_KEY } from "../config/env";
 import { SendInBlueApiError } from "../config/errors";
 import { render } from "../services/renderer";
 
-type RemoteTemplateSlug =
+export type RemoteTemplateSlug =
   | "join-organization"
   | "verify-email"
   | "reset-password"
   | "magic-link"
   | "choose-sponsor"
   | "official-contact-email-verification";
-type LocalTemplateSlug =
+export type LocalTemplateSlug =
   | "organization-welcome"
   | "unable-to-auto-join-organization"
   | "unable-to-find-sponsor"
@@ -39,57 +35,6 @@ const hasRemoteTemplate = (
   template: RemoteTemplateSlug | LocalTemplateSlug,
 ): template is RemoteTemplateSlug =>
   remoteTemplateSlugToSendinblueTemplateId.hasOwnProperty(template);
-
-export async function sendZammadMail({
-  to = [],
-  cc = [],
-  subject,
-  template,
-  params,
-  senderEmail = "moncomptepro@beta.gouv.fr",
-}: {
-  to: string[];
-  cc?: string[];
-  subject: string;
-  template: RemoteTemplateSlug | LocalTemplateSlug;
-  params: any;
-  senderEmail?: string;
-}) {
-  const CREATE_TICKET_ENDPOINT = `${ZAMMAD_URL}/api/v1/tickets`;
-  console.log("Sending mail to Zammad ", CREATE_TICKET_ENDPOINT);
-
-  const data = {
-    cc: undefined as { email: string }[] | undefined,
-    sender: {
-      name: "L’équipe MonComptePro",
-      email: senderEmail,
-    },
-    replyTo: {
-      name: "L’équipe MonComptePro",
-      email: senderEmail,
-    },
-    // Sendinblue allow a maximum of 99 recipients
-    to: chain(to)
-      .sampleSize(99)
-      .map((e) => ({ email: e }))
-      .value(),
-    subject,
-    params,
-    tags: [template],
-    headers: {
-      charset: "iso-8859-1",
-    },
-    templateId: 0,
-  };
-
-  if (DO_NOT_SEND_MAIL) {
-    console.log(`${template} mail not send to ${to}:`);
-    console.log(data);
-    return;
-  }
-
-  return Promise.reject("Not implemented");
-}
 
 export const sendMail = async ({
   to = [],
